Mimic (#15)
* Add tests for training with the mimic dataset

* train with mimic dataset

* relock

* remove psycopg2

* silence some preprocessing warning and bind the file path list to the lambda to avoid reusing the same data every time

* more lambda variable binding

* user proper frequency

* Add libclang to fix a poetry bug, refactor the split window preprocessing and add a filter if the recording has no valid windows

* drop the batch remainders in the dataset

* mount data for training with mimic files in container

* run slurm interactively

* revert to sbatch and increase training time

* remove unused file and train with the full dataset

* restore configuration.py

* remove more __pycache__

* remove more __pycache__

* increase to 100 epochs and reduce to 10000 steps per epochs

* increase training time to 1 day

* log pid

* Fix typo

* reduce the amount of compute resources and training time

* fix a package name

* increase memory during training

* Mimic (#13)

* Add tests for training with the mimic dataset

* train with mimic dataset

* relock

* remove psycopg2

* silence some preprocessing warning and bind the file path list to the lambda to avoid reusing the same data every time

* more lambda variable binding

* user proper frequency

* Add libclang to fix a poetry bug, refactor the split window preprocessing and add a filter if the recording has no valid windows

* drop the batch remainders in the dataset

* mount data for training with mimic files in container

* run slurm interactively

* revert to sbatch and increase training time

* remove unused file and train with the full dataset

* restore configuration.py

* remove more __pycache__

* remove more __pycache__

* increase to 100 epochs and reduce to 10000 steps per epochs

* increase training time to 1 day

* log pid

* Fix typo

* reduce the amount of compute resources and training time

* swithc to tensorflow-gpu as base image and get rid of keras imports

* switch to a gpu enabled tensorflow image and change the python 3.9

* downgrade to python 3.8 to simplify the docker build steps

* train the whole dataset per epoch to avoid run out of data warnings

* Implement some optimization in the preprocessing to speedup training (#14)

* Mimic (#13)

* Add tests for training with the mimic dataset

* train with mimic dataset

* relock

* remove psycopg2

* silence some preprocessing warning and bind the file path list to the lambda to avoid reusing the same data every time

* more lambda variable binding

* user proper frequency

* Add libclang to fix a poetry bug, refactor the split window preprocessing and add a filter if the recording has no valid windows

* drop the batch remainders in the dataset

* mount data for training with mimic files in container

* run slurm interactively

* revert to sbatch and increase training time

* remove unused file and train with the full dataset

* restore configuration.py

* remove more __pycache__

* remove more __pycache__

* increase to 100 epochs and reduce to 10000 steps per epochs

* increase training time to 1 day

* log pid

* Fix typo

* reduce the amount of compute resources and training time

* Use interleave for the datasets

* shuffle the tracks and prefetch the inputs of the model

* reduce training time

* use float32 to speed up training

* convert the signal filter to a numpy array of float32 before returning

* use correct dataset and increase training time to 2 days

* use more CPU cores for training

* train for a week

* reduce CPU count and train time

* remove train.sh from docekr image

+++ tests/preprocessing/pipelines/heartbeatpreprocessing.py
@@ -1,7 +1,7 @@
 from unittest import TestCase
 
 from neurokit2 import ppg_simulate
-from tensorflow import TensorSpec, float64
+from tensorflow import TensorSpec, float32
 from tensorflow.python.data import Dataset
 
 from mlbpestimation.preprocessing.pipelines.heartbeatpreprocessing import HeartbeatPreprocessing
@@ -16,7 +16,7 @@ class TestHeartbeatPreprocessing(TestCase):
 
         processed_dataset = pipeline.preprocess(dataset)
 
-        expected_specs = (TensorSpec(shape=400, dtype=float64), TensorSpec(shape=2, dtype=float64))
+        expected_specs = (TensorSpec(shape=400, dtype=float32), TensorSpec(shape=2, dtype=float32))
         self.assertEqual(expected_specs, processed_dataset.element_spec)
 
     def test_has_data(self):
